refactor(suite-native): move ExpoRouterNavigationBridge into navigation package
The bridge does navigation analytics + sentry breadcrumb tracking and
owns IsNavigationReadyContext — it belongs next to the rest of the
navigation infrastructure, not inline in the app entry. Moving it back
lets useReportSendFlowExitToAnalytics stop being re-exported from the
navigation package's public API (it was only exposed so the inline
bridge could reach it from the app package).

_layout.tsx loses ~80 lines of inlined bridge code plus 6 now-unused
imports; the navigation package gains one focused component file.
