+++ suite-native/navigation/src/index.ts
@@ -15,8 +15,8 @@ export * from './components/ScreenHeader';
 export * from './components/DynamicHeader/DynamicScreenHeader';
 export * from './components/DynamicHeader/DynamicScreenHeaderContext';
 export * from './components/LoadingSuccessScreen';
+export * from './components/ExpoRouterNavigationBridge';
 export * from './components/NavigationContainerWithAnalytics';
 export * from './components/GoBackIcon';
 export * from './hooks/useNavigationRouteMatch';
 export * from './hooks/useDisableIOSGesture';
-export * from './hooks/useReportSendFlowExitToAnalytics';
